[MISC] Rename header files and update includes
